Split tests in two files, use white-box testing
